refactor(agent): fold @agent/core/definition deep-imports behind a definition barrel
Final #10024 cluster: migrate the CLI/desktop/extension
AgentConfig deep imports onto a curated src/agent/core/definition
barrel (AgentConfigSchema, AgentConfig, AgentConfigPayload) and swap
the host-agent baseline specifier 1-for-1 with no widening.

Closes #10024
