+++ packages/desktop/src/main/desktopAgentExecution.ts
@@ -19,7 +19,7 @@ import {
   type SessionHostInteractions,
 } from '@agent/runtime';
 import { presentFollowUpResult, submitFollowUp } from '@agent/followUp';
-import type { AgentConfig } from '@agent/core/definition/AgentConfig';
+import type { AgentConfig } from '@agent/core/definition';
 import {
   validateExecutionRequest,
   type ExecutionRequest,
