+++ packages/desktop/src/main/desktopHistoryHandlers.ts
@@ -1,7 +1,7 @@
 import { readFileSync } from 'node:fs';
 import * as path from 'node:path';
 
-import type { AgentConfig } from '@agent/core/definition/AgentConfig';
+import type { AgentConfig } from '@agent/core/definition';
 import type { ValidatedExecutionRequest } from '@agent/core/state/executionRequests';
 import type { ChatExportController } from '@controllers/settingsView/ChatExportController';
 import {
