+++ packages/extension/src/commands/setup/setupAssistantCommand.ts
@@ -4,7 +4,7 @@ import * as vscode from 'vscode';
 // Local imports
 import { loadAgents } from '@agent/index';
 import { defaultSession, runAgent } from '@agent/runtime';
-import { AgentConfigSchema } from '@agent/core/definition/AgentConfig';
+import { AgentConfigSchema } from '@agent/core/definition';
 import { AUTH_COMMANDS } from '@auth/constants';
 import { EXTENSION_COMMANDS } from '@commands/extensionCommandIds';
 import { apiKeyCommands } from '@commands/api/apiKeyCommands';
